refactor: 삼항조건 연산자 수정, 아이콘 이름 수정
